requirements.txt
PyJWT==1.7.1
fastapi==0.61.1
uvicorn==0.11.8
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
passlib==1.7.2
